Remove metatable from env
There's no more __index. You need to explicitly access _G which stores
gmod's environment. This might be changed even to be more explicit
later, like local _G = Autorun.getGlobals()

Also it's a little problematic that it's using _G so you wouldn't be
able to easily get the globals of the env itself, but whatever.

This is part of the initiative to make Autorun-ng as non-detectable as
possible OOTB, even for noobs that might not be super careful.