WIP
Event system, new file structure, better print

Getting crashes with invalid file descriptors, assuming the issue is
something about storing the plugin as a lightuserdata and it becoming
invalidated as the plugin changes.

I think I'm going to have to create a userdata wrapper and store the
plugin objects in lua's userdata allocator, so that they can be properly
stored as upvalues that can be referred to by the callbacks properly.